feature(GameActions): top level class for actions between states
